// src/material-design/colors-type.ts
export type Colors = {
  primary: string
  onPrimary: string
  primaryContainer: string
  onPrimaryContainer: string
  inversePrimary: string

  primaryFixed: string
  primaryFixedDim: string
  onPrimaryFixed: string
  onPrimaryFixedVariant: string

  secondary: string
  onSecondary: string
  secondaryContainer: string
  onSecondaryContainer: string

  secondaryFixed: string
  secondaryFixedDim: string
  onSecondaryFixed: string
  onSecondaryFixedVariant: string

  tertiary: string
  onTertiary: string
  tertiaryContainer: string
  onTertiaryContainer: string

  tertiaryFixed: string
  tertiaryFixedDim: string
  onTertiaryFixed: string
  onTertiaryFixedVariant: string

  error: string
  onError: string
  errorContainer: string
  onErrorContainer: string

  background: string
  onBackground: string

  surface: string
  onSurface: string
  surfaceVariant: string
  onSurfaceVariant: string
  inverseSurface: string
  inverseOnSurface: string
  surfaceTint: string
  surfaceTintColor: string

  surfaceDim: string
  surfaceBright: string
  surfaceContainerLowest: string
  surfaceContainerLow: string
  surfaceContainer: string
  surfaceContainerHigh: string
  surfaceContainerHighest: string

  outline: string
  outlineVariant: string

  shadow: string
  scrim: string

  surfaceDisabled: string

  elevation0: string
  elevation1: string
  elevation2: string
  elevation3: string
  elevation4: string
  elevation5: string
}